+++ CHANGES.md
@@ -12,6 +12,11 @@ Since 0.53.0, we officially provide
 `Libplanet.KeyStore.Web3KeyStore`) and complies with the `KeyStore` interface
 defined in *[@planetarium/account]*.
 
+Due to changes in [[#2961]], a network ran with [Libplanet 0.51.0] or
+[Libplanet 0.52.0] is not compatible with this version.  Thus, it is
+recommended to skip [Libplanet 0.51.0] and [Libplanet 0.52.0] for
+deployment if possible.
+
 ### Backward-incompatible API changes
 
  -  Removed state completion feature with related classes as below.  [[#2703]]
@@ -57,6 +62,14 @@ defined in *[@planetarium/account]*.
  -  Changed `BlockHash` to implement `IBencodable`.  [[#2959]]
      -  Changed `BlockHash(Binary)` to `BlockHash(IValue)`.
      -  Added `BlockHash.Bencoded` property.
+ -  (Libplanet.Net) Changed `VoteMetadata` and `Vote` to implement
+    `IBencodable`.  [[#2961]]
+     -  Changed `VoteMetadata(Dictionary)` to `VoteMetadata(IValue)`
+        and `Vote(Dictionary)` to `Vote(IValue)`.
+     -  Changed `VoteMetadata.Encoded` to `VoteMetadata.Bencoded`
+        and `Vote.Encoded` to `Vote.Bencoded`.
+ -  (Libplanet.Net) Changed `BlockCommit`'s `IValue` encoding format.
+    [[#2961]]
 
 ### Backward-incompatible network protocol changes
 
@@ -84,11 +97,13 @@ defined in *[@planetarium/account]*.
 
 ### CLI tools
 
+[Libplanet 0.52.0]: https://www.nuget.org/packages/Libplanet/0.52.0
 [#2703]: https://github.com/planetarium/libplanet/pull/2703
 [#2915]: https://github.com/planetarium/libplanet/pull/2915
 [#2954]: https://github.com/planetarium/libplanet/pull/2954
 [#2955]: https://github.com/planetarium/libplanet/pull/2955
 [#2959]: https://github.com/planetarium/libplanet/pull/2959
+[#2961]: https://github.com/planetarium/libplanet/pull/2961
 [@planetarium/account-web3-secret-storage]: https://www.npmjs.com/package/@planetarium/account-web3-secret-storage
 [@planetarium/bencodex]: https://www.npmjs.com/package/@planetarium/bencodex
 [disjukr/bencodex]: https://github.com/disjukr/bencodex